rename: move NotFound from components to page (#186)
